refactor targets: sdl_runtime&viewer
