+++ test/render/render.cpp
@@ -1,7 +1,7 @@
 #include "Application/Event/Event.h"
 #include "Application/AppBase.hpp"
 #include "TestEventListener.h"
-#include "Entry/SDL/SDLImpl/AppSDLImpl.hpp"
+#include "Entry/SDL/AppSDLImpl.hpp"
 #ifdef VGG_USE_VULKAN
 #include "Entry/SDL/SDLImpl/AppSDLVkImpl.hpp"
 #endif
